Fixes for implicit conversion warnings JIRA CBLT-189
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100474 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/cobalt/seq.hpp
@@ -122,7 +122,7 @@ public:
     /// Get the length of the current sequence
     /// @return Sequence length
     ///
-    int GetLength() const { return m_Sequence.size(); }
+    int GetLength() const { return (int)m_Sequence.size(); }
 
     /// Given an edit script, insert gaps into a sequence. The
     /// profile of the sequence is adjusted automatically, and
